moved file reference conversion to it's own file, fixed imports of shared models, and fixed file_name -> file_path as per model change for file references
