+++ bia-ingest-shared-models/bia_ingest_sm/conversion/image_acquisition.py
@@ -10,7 +10,7 @@ from ..biostudies import (
     Submission,
     attributes_to_dict,
 )
-from src.bia_models import bia_data_model
+from bia_shared_datamodels import bia_data_model
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO)
