+++ bia-ingest-shared-models/bia_ingest_sm/conversion/study.py
@@ -14,7 +14,7 @@ from ..biostudies import (
     attributes_to_dict,
 )
 from ..config import settings
-from src.bia_models import bia_data_model, semantic_models
+from bia_shared_datamodels import bia_data_model, semantic_models
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO)
