+++ bia-ingest-shared-models/bia_ingest_sm/conversion/utils.py
@@ -1,6 +1,5 @@
 import logging
 from pathlib import Path
-import re
 import hashlib
 import uuid
 from typing import List, Any, Dict, Optional, Tuple, Type, Union
@@ -10,12 +9,8 @@ from ..biostudies import (
     attributes_to_dict,
     Section,
     Attribute,
-    find_file_lists_in_submission,
-    flist_from_flist_fname,
-    file_uri,
 )
 from ..config import settings
-from src.bia_models import bia_data_model, semantic_models
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO)
